+++ client/src/components/CustomCategoriesSettingsModal/CustomCategoriesSettingsModal.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { SavedItemsSettingsModal } from './SavedItemsSettingsModal';
+import { SavedItemsSettingsModal } from '../SavedItemSettingsModal/SavedItemsSettingsModal';
 import { useSelector, useDispatch } from 'react-redux';
-import { selectModalOpen, selectData, deleteCategory } from '../store/slices/dataSlice';
-import { StoreCategory } from '../models/Store';
-import CustomCategoriesSettingItem from './CustomCategoriesSettingItem';
+import { selectModalOpen, selectData, deleteCategory } from '../../store/slices/dataSlice';
+import { StoreCategory } from '../../models/Store';
+import CustomCategoriesSettingItem from '../CustomCategoriesSettingsItem/CustomCategoriesSettingItem';
 import { Modal, IconButton } from '@mui/material';
-import AddCategoryPopOver  from './AddCategoryPopOver/AddCategoryPopOver';
+import AddCategoryPopOver  from '../AddCategoryPopOver/AddCategoryPopOver';
 import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 
 export const CustomCategoriesSettingsModal = ({
