+++ client/src/components/CustomCategoryPanel/CustomCategoryPanel.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { Category } from './CustomCategory';
-import { StoreCategory } from '../models/Store';
+import { Category } from '../CustomCategory/CustomCategory';
+import { StoreCategory } from '../../models/Store';
 import IconButton from '@mui/material/IconButton';
 import SettingsIcon from '@mui/icons-material/Settings';
-import { toggleSystemCategorySettings } from '../store/slices/dataSlice';
+import { toggleSystemCategorySettings } from '../../store/slices/dataSlice';
 import { useDispatch } from 'react-redux';
 const iconSx = { width: 18, height: 18, marginLeft: '-10px', marginTop: '-10px', color: 'rgba(255, 255, 255, 0.503)' };
 const warningStyle = {color: 'white',
