+++ client/src/components/SquareItem/SquareItem.tsx
@@ -3,7 +3,7 @@ import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
-import { SavedItem } from '../models/SavedItem';
+import { SavedItem } from '../../models/SavedItem';
 
 export const Bookmark = ({ item }: { item: SavedItem }) => {
     const [isHovered, setIsHovered] = React.useState(false);
